Split `Module` logic into `InstallableModule` and `ExportableModule`.
